Issue #377 - use the VMUSB drivers in place rather than as copies
Incorporate the marker directl from vmusb/daqconfig rather than
an unmaintainable copy.
