+++ main/mvlc/Makefile.am
@@ -65,11 +65,10 @@ IMPORTED_MODULES=$(IMPORT_DIR)/CMQDC32RdoHdwr.h \
 	$(IMPORT_DIR)/CMesytecBase.cpp $(IMPORT_DIR)/CMesytecBase.h \
 	$(IMPORT_DIR)/CMADC32.cpp $(IMPORT_DIR)/CMADC32.h   \
 	$(IMPORT_DIR)/CMADCChain.h $(IMPORT_DIR)/CMADCChain.cpp  \
-	$(IMPORT_DIR)/CMADCScaler.h $(IMPORT_DIR)/CMADCScaler.cpp
+	$(IMPORT_DIR)/CMADCScaler.h $(IMPORT_DIR)/CMADCScaler.cpp \
+	$(IMPORT_DIR)/CMarker.h $(IMPORT_DIR)/CMarker.cpp 
 
-
-LOCAL_MODULES= CMarker.h CMarker.cpp \
-	MADC32Registers.h \
+LOCAL_MODULES= 	MADC32Registers.h \
 	CMTDC32.h CMTDC32.cpp \
 	CMDPP16QDC.cpp  CMDPP16QDC.h  CMDPP32QDC.cpp  CMDPP32QDC.h \
 	CMDPP32SCP.cpp  CMDPP32SCP.h  CMDPP.h  CMDPPQDC.h \
